🗑️ Remove cast

+++ src/Engine.sol
@@ -263,7 +263,7 @@ contract Engine is IEngine, EIP712, EIP7412 {
     {
         // shift _nonce to the right by 8 bits and cast to uint248
         /// @dev wordPos == 0 if 0 <= _nonce <= 255, 1 if 256 <= _nonce <= 511, etc.
-        wordPos = uint248(_nonce >> 8);
+        wordPos = _nonce >> 8;
 
         // cast the last 8 bits of _nonce to uint8
         /// @dev 0 <= bitPos <= 255
